Intermediary commit showing fetch from userprofile feed
The mobile app's home page uses a user profile's feed as the source of data. The feed can contain events or posts. If it does contain an event, the true event data then has to be fetched from the events collection.

This is in an intermediary commit whose follow-up will be fetching the events.

The database design is not agreeable.